Update customizations.en.md
I'm not sure if this is the best place to make this proposal but I think the repo is private. It'll be great if you could change the document number input to 'type=number' in your Checkout Pro form. Right now it have 'type=text' and when you open it from a Webview (app mobile - react native) is not the best experience.
https://user-images.githubusercontent.com/40902023/148293616-4ee3d73f-2daf-4190-9ad7-8f67dc8a413c.png